Document how to create a rootfs and run the kernel
* detailed instructions how to create an inital rootfs based on
busybox
* how to use the rootfs as initramfs
* commands for U-Boot how to load and run the Kernel

Instructions for beginners to give a howto guide to run a self-compiled
kernel on the OrangePi5Plus board. Only original upstream repositories
get used to keep board vendor impact as low as possible.

The goal is to end up a TTY shell of the self-compiled busybox.

Adds a troubleshoot section with some errors that occurred on my own
journey to get the kernel running on an OrangePi 5 Plus board.

Maybe this helps others to get their kernel working.