Add main Makefile to build a rootfs
The Makefile clones the buildroot repo to use buildroot to build the
rootfs. The configuration for buildroot and some additional packages
will be added dynamically. Furthermore it is planned to check the hashes
of the packages again in case buildroot verifies against hashes that are
not original. This way, it should get visible if any package has changed
outside of our repository e.g. in case of an update of buildroot.

buildroot applies also patches to some packages. The patches need to get
reviewed.