Verify buildroots git tag hash and expose patches
Verifies the hash of the defined buildroot tag to be sure the tag got
not changed.

Adds a proxy-script that runs instead of the patch binary to expose the
piped in patches before the actual patch gets applied. Since buildroot
runs patches not only from the packages/-directories but also from
support/*/*.patch, we see every patch that gets applied via the
patch-bin, also the hidden ones.

This enables us to see all applied patches during build and review them
and makes it better possible to discover possible backdoors that got
introduced via patches.