Enable RTL8125 Ethernet over PCIe
There is no functional ethernet device on the Opi5plus board. Instead,
the 2.5Gbp ports are connected via PCIe and use a RTL8125BG device.
Because of of that, the ethernet devices (gmac0/gmac1) are disabled in
the device tree and the RTL8125 drivers need to be enabled in the kernel
config.