Add script to create rootfs
The rootfs script copies all required files into rootfs directories and
creates a rootfs image out of them. The rootfs contains the kernel files
and u-boot.scr with u-boot commands to boot the specified kernel and the
initramfs.

The initramfs performs a switch_root to the mounted rootfs. It falls
back into a shell on failure.